rename sdk/auditlog root package from log to auditlog
