+++ sdk/auditlog/audit_errors.go
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package log
+package auditlog
 
 import "go.opentelemetry.io/otel/sdk/auditlog/status"
