+++ sdk/auditlog/audit_logger_api.go
@@ -1,7 +1,7 @@
 // Copyright The OpenTelemetry Authors
 // SPDX-License-Identifier: Apache-2.0
 
-package log
+package auditlog
 
 import (
 	"context"
